+++ cmd/ssql-playground/main.go
@@ -158,6 +158,7 @@ func buildCommand() *cf.Command {
 
 	cmd = commands.RegisterVersion(cmd)
 	cmd = commands.RegisterFunctions(cmd)
+	cmd = commands.RegisterConventions(cmd)
 	cmd = commands.RegisterFrom(cmd)
 	cmd = commands.RegisterLimit(cmd)
 	cmd = commands.RegisterOffset(cmd)